✨ Add blending utilities for core and edge profile functions
Separate blending function from plasma module and move it into math module
